Decode a full VDP_VIDEO_MIXER_PICTURE_STRUCTURE_FRAME
When the original version of this program was written,
it was necessary to "lie" about an H.265/HEVC frame, and
describe it as a top field or bottom field. This was due
to a bug in the NVIDIA VDPAU implementation. This bug has
since been fixed. So, declare the output frame to be a
VDP_VIDEO_MIXER_PICTURE_STRUCTURE_FRAME.